requirements.txt
Flask
Flask-MySQLdb
Flask-SQLAlchemy
requests
python-dotenv
pytest
